requirements.txt
Django==4.2.11
daphne==4.1.2
channels==4.0.0
channels-redis==4.2.0
djangorestframework==3.15.1
django-cors-headers==4.4.0
python-decouple==3.8
requests==2.32.3
